# Speed up vec256 bfloat16 type conversion on aarch64