This PR speeds up vec256 bfloat16 type conversion on aarch64. Test AOTI compile with gpt-fast/generate.py (https://github.com/pytorch-labs/gpt-fast) stories15M model on Apple M2. The perf improves from 187 tokens/sec to 330 tokens/sec. Raw data as below:

Before:
```
(bf16) helloguo@helloguo-mac gpt-fast % python generate.py --checkpoint_path checkpoints/stories15M/stories15M.pt --prompt "Once upon a time" --device cpu --num_samples 20 --compile
Using device=cpu
Loading model ...
Time to load model: 0.04 seconds
/Users/helloguo/Desktop/local_llm/pytorch/torch/backends/cuda/__init__.py:393: FutureWarning: torch.backends.cuda.sdp_kernel() is deprecated. In the future, this context manager will be removed. Please see, torch.nn.attention.sdpa_kernel() for the new context manager, with updated signature.
  warnings.warn(
/Users/helloguo/Desktop/local_llm/pytorch/torch/backends/cuda/__init__.py:393: FutureWarning: torch.backends.cuda.sdp_kernel() is deprecated. In the future, this context manager will be removed. Please see, torch.nn.attention.sdpa_kernel() for the new context manager, with updated signature.
  warnings.warn(
Compilation time: 9.56 seconds
Once upon a time, there was a little girl named Lily. She had a teddy bear named Teddy. Teddy belonged to her fairy, which made her very happy.
One day, Lily spilled her juice on Teddy and made him wet. She didn't know what to do, so she went to her mommy and told her what happened. Her mommy said, "Don't worry, Lily. We can clean Teddy together."
Lily and her mommy took Teddy to a dry place and washed him with soap and water. Teddy was all clean again! Lily was so happy and hugged her mommy to say thank you. From that day on, Lily made sure to keep her toys safe and dry. Once upon a time, there was a little girl named Lily. She loved to play with her toys and run around outside. One day, she saw a big, scary bear
Time for inference 1: 1.08 sec total, 184.65 tokens/sec
Bandwidth achieved: 9.01 GB/s
Once upon a time, there was a little girl named Lily. She loved to play outside in the sunshine. One day, she went to the park with her mommy. They played on the swings and went down the slide.
Suddenly, Lily saw something shiny on the ground. She ran over to it and found a pretty necklace! She was so happy and showed her mommy. "Look, mommy! I found a pretty necklace!" she said.
After playing for a while, Lily and her mommy went back home. But when they got there, they saw that the sun was setting. "Oh no! We're going to get lost," said Lily's mommy. "We need to find our way back home."
Lily felt bitter because she wanted to keep playing, but she knew it was important to listen to her mommy. They walked back home, holding hands and talking about all the fun things
Time for inference 2: 1.03 sec total, 193.73 tokens/sec
Bandwidth achieved: 9.46 GB/s
Once upon a time, there was a little boy named Timmy. He loved to play outside and explore the forest near his house. One day, as he was walking, he saw a big, scary bear! The bear started to chase him and Timmy was very scared. 
But then, Timmy remembered what his mom had told him: "If you ever get scared, just run away from danger." So, he ran as fast as he could back to his house. When he got there, he told his mom what had happened. She hugged him and said, "Don't worry, everything will be okay." 
From then on, Timmy was always careful when he played outside. He knew that if he ever felt scared, he could always run away from danger. And he always remembered to be careful and look both ways before he went anywhere. Once upon a time, there was a little girl named Lily. She loved to play outside
Time for inference 3: 1.03 sec total, 194.22 tokens/sec
Bandwidth achieved: 9.48 GB/s
Once upon a time, there was a little girl named Lily who loved to wear colorful dresses. One day, she saw a beautiful bracelet in a store window and wanted it so badly. She asked her mommy if she could have it, but her mommy said it cost too much money.
Lily was sad and went to her room to cry. Suddenly, she heard a knock on her door. It was her friend Timmy who had a toy car made of gold. Lily was so happy and asked Timmy to load it into her wagon to save some money.
As Timmy was loading the car, Lily saw a little boy who looked sad. She asked him what was wrong and he said he lost his toy and couldn't find it. Lily remembered that she had a pretty bracelet that she saved up and decided to buy him a new toy car. Timmy was so happy and grateful.
But then,
Time for inference 4: 1.03 sec total, 194.81 tokens/sec
Bandwidth achieved: 9.51 GB/s
Once upon a time, there was a little girl named Lily. She had a toy box that she loved to play with. One day, her mom told her to comb her hair because it was dirty from playing outside. 
Lily tried to comb her hair, but it was hard to do. She asked her friend, a small bird, to help her. The bird sat on her shoulder and started to comb Lily's hair. 
While the bird was combing, something unexpected happened. The bird flew into Lily's room and landed on her head! Lily was surprised and didn't know what to do. The bird was very smart and explained that she had just been practicing to tie her hair back into a bird outfit. 
Lily was happy to have a new pet bird, and she allowed the bird to stay for her study. They became the best of friends, and Lily learned that sometimes unexpected things can happen, but
Time for inference 5: 1.05 sec total, 190.65 tokens/sec
Bandwidth achieved: 9.31 GB/s
Once upon a time, there was a little girl named Lily. She loved her blanket very much and always slept with it at night. One day, Lily's mom asked her if she wanted to take a shower. Lily said yes and ran to get her blanket.
When Lily came back, she saw that her blanket was gone. She asked her mom where it went, but her mom didn't know either. Lily started to get angry and said, "Mommy, where is my blanket? I want it back!"
Her mom replied, "I don't know, Lily. Maybe it's in the washing machine."
Lily was very upset and didn't know what to do. She didn't want to lose her favorite blanket, but she also didn't want to get rid of it. Suddenly, her little brother came into the room and said, "I found your blank
Time for inference 6: 1.07 sec total, 187.32 tokens/sec
Bandwidth achieved: 9.14 GB/s
Once upon a time, there was a little boy named Timmy. Timmy loved to play with his toys all day long. One day, Timmy's mommy took him to the park to play. Timmy saw a big slide and wanted to climb it. He was too clumsy and fell off.
Later, Timmy saw a man playing jazz music and danced around. The man was wearing a guitar and playing jazz music. Timmy thought it was cool and wanted to dance too. He asked the man if he could show him how to play some jazz. The man said yes and Timmy tried to play a saxophone, but he kept falling over the whole time.
As the sun started to set, Timmy and his mommy went home. Timmy was tired from all the running and playing. He fell asleep right away. The next day, Timmy's mommy took him to the park again. This time, Tim
Time for inference 7: 1.03 sec total, 193.28 tokens/sec
Bandwidth achieved: 9.44 GB/s
Once upon a time, there was a little boy named Tim. He loved to play outside in the park. One day, he found a big, red ball. Tim was very happy and played with the ball all day.
After playing, Tim felt hungry. He went to his mom and said, "Mom, I am hungry. Can you feed me?" His mom gave him a yummy ice cream to eat. Tim ate the ice cream fast, but he ate his ice cream fast too.
Tim and his mom went home. Tim told his dad about the yummy ice cream he ate. His dad was happy because the ice cream made the day even better. Tim felt good that he could eat all the ice cream that morning. One day, a little boy named Tim was playing with his toys. He had a big box full of toys and a small table. He was very happy. He wanted to play with
Time for inference 8: 1.04 sec total, 192.35 tokens/sec
Bandwidth achieved: 9.39 GB/s
Once upon a time, in a big, green garden, there was a pretty poppy. The poppy had many friends who played with it every day. They liked to jump and dance in the sun. The poppy was very happy.
One day, a big, mean bird came to the garden. The bird wanted to eat the pretty poppy and the other flowers. But the poppy was very nice and wanted to protect the flowers and its friends.
The poppy had an idea. It went to the big, mean bird and said, "You cannot eat our pretty poppy and your friends. We will protect them and keep the garden safe." The big, mean bird was surprised. It did not want to hurt the poppy and its friends. So, the bird flew away, and the poppy and the flowers were safe. They were very happy and continued to play in the big, green garden. One day, a little dog named Max was in his house. He
Time for inference 9: 1.06 sec total, 188.59 tokens/sec
Bandwidth achieved: 9.21 GB/s
Once upon a time, there was a little girl named Lily. She loved to play in the park with her friends. One day, she saw a big fountain in the middle of the park. It was so pretty with the water shining in the sun.
Lily's friend, Timmy, said, "Let's go play in the fountain!" But Lily said, "No, Timmy. It's too cold and the fountain is not fun to play in."
Timmy didn't listen and went into the fountain. He was very cold and started to shiver. Lily called out to him, "Timmy, come back! It's too cold in here!" But it was too late. Timmy got caught in the fountain and got very cold and sick. He had to go to the hospital and couldn't play in the park anymore.
Lily felt very sad and wished that Timmy
Time for inference 10: 1.08 sec total, 185.02 tokens/sec
Bandwidth achieved: 9.03 GB/s
Once upon a time, there was a famous cat named Tom. Tom was not like other cats. He could talk like people. He loved to help people.
One day, Tom saw a little girl named Sue. Sue was sad because her cat was stuck in a tree. "Help me, please!" she cried. Tom said, "Don't worry, I will help you." Tom went to the tree and used his strong muscles to help Sue get her cat down.
Sue was very happy and said, "Thank you, Tom!" Tom felt good helping Sue. He learned that helping others is a good thing to do. From that day on, Tom tried to be kind and helpful to everyone he met. Once upon a time, there was a little girl named Lily. She had a sister named Emily who was always naughty. One day, Emily wanted to play with Lily's favorite toy, but Lily didn't want to share
Time for inference 11: 1.08 sec total, 184.55 tokens/sec
Bandwidth achieved: 9.01 GB/s
Once upon a time, there was a little boy named Timmy. He loved to play with his toy cars and trucks. One day, he was playing with his red truck when he accidentally broke it.
"Oh no!" cried Timmy. "My truck is broken!"
His mom heard his cries and came to see what happened. "What happened, Timmy?" she asked.
"I broke my truck," Timmy said sadly.
"It's okay," said his mom. "I'll fix it for you."
After his mom fixed the truck, she noticed that some of the wheels were missing. She took them off and put the wheels back on. Timmy was happy again and continued to play with his red truck. Once upon a time, there was a shy dog named Spot. Spot liked to play near a big tree. One day, Spot found a shiny gem
Time for inference 12: 1.06 sec total, 189.01 tokens/sec
Bandwidth achieved: 9.23 GB/s
Once upon a time, there was a little bird named Billy. Billy lived in a high tree with his family. He loved to fly and play in the sky. One day, while Billy was flying, he saw a big cloud. He thought it would be fun to play under the cloud.
Billy said to his mom, "Can I play under the big cloud?" His mom said, "No, Billy. It's too high and cold." But Billy did not listen. He flew closer to the big cloud and got very, very cold. He started to shiver.
Just then, a kind bird named Betty saw Billy shivering. Betty flew down and said, "Billy, I warn you, don't play there. It's not safe." Billy listened to Betty and flew back to his high tree. He was happy to be safe and warm inside its warm home. Once upon a time, there was a little girl named Lily. She
Time for inference 13: 1.06 sec total, 188.94 tokens/sec
Bandwidth achieved: 9.22 GB/s
Once upon a time, there was a boy named Timmy. Timmy loved to play with his toys. One day, he found a big, shiny missile in his toy box. He was gifted and could do lots of cool things. 
Timmy wanted to play with the missile but he didn't know how to use it. So, he asked his mom to help him. His mom showed him how to use the missile and they had lots of fun playing with it. 
After a while, Timmy wanted to throw the missile away, but his mom told him that it was not a toy and should be left alone. So, Timmy put the missile away and promised to play with it in a safe place. Once upon a time, there was a little girl named Lily. One day, Lily found a bottle of perfume in the park. She thought it smelled good, so she picked it up and sm
Time for inference 14: 1.07 sec total, 187.69 tokens/sec
Bandwidth achieved: 9.16 GB/s
Once upon a time, there was a little boy named Timmy. Timmy loved playing with his toy cars and trucks. One day, Timmy's dad brought home a big truck that could carry load many things.
Timmy was so excited to play with the truck. He asked his dad, "Can I carry the truck with me?" His dad replied, "Sure, but be careful. It's very strong."
Timmy held the truck up high and started to push it around. But then, the truck started to wobble and Timmy got frustrated. "I can't do it!" he said.
His dad said, "That's okay, Timmy. You can try again. Just be more careful." Timmy tried again and this time he was able to march around the truck without getting frustrated. "I did it!" he exclaimed. His dad smiled and
Time for inference 15: 1.08 sec total, 184.64 tokens/sec
Bandwidth achieved: 9.01 GB/s
Once upon a time, in a small village, there was a little girl named Lily. Lily had a big, red ball. She liked to play with her ball all day long. One day, she saw a small, scared cat near her house. The cat was hiding in the bushes.
Lily went to the cat and said, "Don't be scared, little cat. I will help you." She picked up her ball and gave it to the cat. The cat was very happy and started to play with the ball.
As they played, Lily saw a boy who was sad. She wanted to make him happy. She thought, "I can offer him my ball." So, Lily asked the boy, "Do you want to play with my ball?" The boy smiled and said, "Yes, please!" They played together and became good friends.
The moral of the story is, when you make others happy, you can find true friends by being
Time for inference 16: 1.13 sec total, 176.45 tokens/sec
Bandwidth achieved: 8.61 GB/s
Once upon a time, there was a little girl named Lily. She loved to play outside with her friends. One day, Lily and her friends were playing a game of tag. Lily was running so fast that she slipped on a banana peel and fell down. 
After a few minutes, Lily's knee felt better and she started to skip again. Her friends joined in and helped her up. They all skipped together, laughing and having fun. 
When it was time to go home, Lily's mom had a surprise for her. She had bought Lily a new toy! Lily was so eager to play with it that she forgot all about her hurt knee. She ran to the toy and started playing with it. She had so much fun that she forgot all about her pain and the toy. From that day on, Lily was always careful when playing outside so she wouldn't slip again. One
Time for inference 17: 1.07 sec total, 186.20 tokens/sec
Bandwidth achieved: 9.09 GB/s
Once upon a time, a little duck named Ducky lived near a big pond. Ducky loved to swim and play in the water. One day, it rained a lot, and the water started to pour. Ducky was scared, but his mommy told him to come inside and remain safe.
Ducky's mommy gave him a bottle of water to drink. Ducky drank the water and felt better. He looked at the pond and saw the water was dry. Ducky was happy that he could remain safe and warm.
The sun came out and dried up all the water. Ducky's mommy was proud of him for being brave and staying safe. She gave him a big hug and said, "Good job, Ducky!" Ducky felt happy and loved. He knew that he was safe and loved, and he would always remain close to his mommy. Once upon a time, there was a crab
Time for inference 18: 1.08 sec total, 185.52 tokens/sec
Bandwidth achieved: 9.06 GB/s
Once upon a time, there was a little girl named Lily. She loved to play with her toys and run around in the park. One day, Lily's mommy told her they were going to the store to buy a new uniform for a special party. Lily was excited to wear her new uniform because she knew it was dependable and would keep her best.
On the day of the party, Lily put on her new uniform and felt like a real princess. But when they got to the party, Lily realized she had lost her favorite toy. She searched everywhere, but she couldn't find it. She started to cry and her mommy tried to comfort her.
Suddenly, Lily's little brother came up to her and offered to play with her toys. He had so much fun racing around the room and waving his toy. Lily realized that losing her toy wasn't so bad after all. She smiled
Time for inference 19: 1.09 sec total, 183.32 tokens/sec
Bandwidth achieved: 8.95 GB/s
Once upon a time, there was a little girl named Lily. She loved cake very much. One day, she made a big, yummy cake for her mommy's birthday. She was very excited to give it to her.
Lily's mommy saw the cake and said, "Lily, you are so good at making cake. Do you want to help me bake a cake?" Lily was so happy and said, "Yes, please!"
Lily and her mommy baked a delicious cake. They put it in the oven and waited. When it was done, they took it out and gave it to Lily's mommy. She said, "Thank you for the cake, Lily. It looks great!" Lily felt very happy and proud of her cake. Once upon a time, there was a little girl named Lily. She loved to play in the park and pick ber
Time for inference 20: 1.08 sec total, 184.46 tokens/sec
Bandwidth achieved: 9.00 GB/s
==========
Average tokens/sec: 187.77
Memory used: 0.00 GB
```
After:
```
(bf16) helloguo@helloguo-mac gpt-fast % python generate.py --checkpoint_path checkpoints/stories15M/stories15M.pt --prompt "Once upon a time" --device cpu --num_samples 20 --compile
Using device=cpu
Loading model ...
Time to load model: 0.02 seconds
/Users/helloguo/Desktop/local_llm/pytorch/torch/backends/cuda/__init__.py:393: FutureWarning: torch.backends.cuda.sdp_kernel() is deprecated. In the future, this context manager will be removed. Please see, torch.nn.attention.sdpa_kernel() for the new context manager, with updated signature.
  warnings.warn(
/Users/helloguo/Desktop/local_llm/pytorch/torch/backends/cuda/__init__.py:393: FutureWarning: torch.backends.cuda.sdp_kernel() is deprecated. In the future, this context manager will be removed. Please see, torch.nn.attention.sdpa_kernel() for the new context manager, with updated signature.
  warnings.warn(
Compilation time: 8.81 seconds
Once upon a time, there was a little girl named Lily. She had a teddy bear named Teddy. Teddy belonged to her fairy, which made her very happy.
One day, Lily spilled her juice on Teddy and made him wet. She didn't know what to do, so she went to her mommy and told her what happened. Her mommy said, "Don't worry, Lily. We can clean Teddy together."
Lily and her mommy took Teddy to a dry place and washed him with soap and water. Teddy was all clean again! Lily was so happy and hugged her mommy to say thank you. From that day on, Lily made sure to keep her toys safe and dry. Once upon a time, there was a little girl named Lily. She loved to play with her toys and run around outside. One day, she saw a big, scary bear
Time for inference 1: 0.59 sec total, 338.10 tokens/sec
Bandwidth achieved: 16.50 GB/s
…